Data scrape
These are most of the companies that I pulled from a some sources.
PhlDesign, Philadelphia Top 100 companies to work for, and NET/WORK job
fair. Companies that had job positions have already been put in. These
are most of the remaining companies. The “underConsideration” companies
have a career url, but not open positions listed. The “needCareerUrl”
companies don’t meet the job page requirement. In any case, it is
possible that some of the companies may never have tech positions and
should be deleted.